requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
cssselect==1.2.0
finviz==1.4.4
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
install==1.3.5
lxml==4.9.2
multidict==6.0.4
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
numpy==1.24.1
pandas==1.5.2
protobuf==3.20.1
pycodestyle==2.10.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
tenacity==7.0.0
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.13
user-agent==0.1.10
yarl==1.8.2
